post

+++ app/Http/Controllers/TasksController.php
@@ -11,11 +11,19 @@ class TasksController extends Controller
     
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            
+            $tasks = $user->tasks()->get();
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
         
-        return view('welcome', [
-            'tasks' => $tasks,
-            ]);
+        return view('welcome', $data);
     }
 
     public function create()
@@ -33,11 +41,11 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required|max:191',
             ]);
-        
-        $task = new Task;
-        $task->status= $request->status;
-        $task->content = $request->content;
-        $task->save();
+    
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
         
         return redirect('/');
     }
@@ -66,19 +74,26 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required|max:191',
             ]);
-        
-        $task = Task::find($id);
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+            
+        $task = \App\Task::find($id);
+            
+        if(\Auth::id() === $task->user_id){
+            $task = Task::find($id);
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
+        }
         
         return redirect('/');
     }
 
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
+        
+        if(\Auth::id() === $task->user_id){
+          $task->delete();
+        }
         
         return redirect('/');
     }
